Update IPv6 port guideline
The linked ipv6 addresses include a port number in them where as the linked ipv4 addresses do not. 

Users following this guide verbatim will encounter a bad source address error if they follow the current guidance like:  ufw insert 2 allow from 2a01:4f8:c17:ed5b::/64/22 to any port 22

For consistency we could consider removing the port number from the ipv6 section here or amending it to be what is in the text file and removing the port from the text file, or just remove the port altogether.